+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalCheckoutRequestUnitTest.java
@@ -5,11 +5,12 @@ import android.os.Parcel;
 import com.braintreepayments.api.core.Authorization;
 import com.braintreepayments.api.core.Configuration;
 import com.braintreepayments.api.core.PostalAddress;
+import com.google.testing.junit.testparameterinjector.TestParameter;
 
 import org.json.JSONException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.robolectric.RobolectricTestRunner;
+import org.robolectric.RobolectricTestParameterInjector;
 
 import java.util.ArrayList;
 
@@ -20,7 +21,7 @@ import static junit.framework.Assert.assertTrue;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Mockito.mock;
 
-@RunWith(RobolectricTestRunner.class)
+@RunWith(RobolectricTestParameterInjector.class)
 public class PayPalCheckoutRequestUnitTest {
 
     @Test
@@ -37,11 +38,13 @@ public class PayPalCheckoutRequestUnitTest {
         assertNull(request.getLandingPageType());
         assertNull(request.getBillingAgreementDescription());
         assertFalse(request.getShouldOfferPayLater());
+        assertFalse(request.getEnablePayPalAppSwitch());
+        assertNull(request.getUserAuthenticationEmail());
         assertFalse(request.getHasUserLocationConsent());
     }
 
     @Test
-    public void setsValuesCorrectly() {
+    public void setsValuesCorrectly(@TestParameter boolean appSwitchEnabled) {
         PostalAddress postalAddress = new PostalAddress();
         PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00", true);
         request.setCurrencyCode("USD");
@@ -56,6 +59,8 @@ public class PayPalCheckoutRequestUnitTest {
         request.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
+        request.setEnablePayPalAppSwitch(appSwitchEnabled);
+        request.setUserAuthenticationEmail("test-email");
         request.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
 
         assertEquals("1.00", request.getAmount());
@@ -69,13 +74,15 @@ public class PayPalCheckoutRequestUnitTest {
         assertEquals(PayPalPaymentUserAction.USER_ACTION_COMMIT, request.getUserAction());
         assertEquals("Display Name", request.getDisplayName());
         assertEquals("123-correlation", request.getRiskCorrelationId());
+        assertEquals(appSwitchEnabled, request.getEnablePayPalAppSwitch());
+        assertEquals("test-email", request.getUserAuthenticationEmail());
         assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
         assertTrue(request.getShouldOfferPayLater());
         assertTrue(request.getHasUserLocationConsent());
     }
 
     @Test
-    public void parcelsCorrectly() {
+    public void parcelsCorrectly(@TestParameter boolean appSwitchEnabled) {
         PayPalCheckoutRequest request = new PayPalCheckoutRequest("12.34", true);
         request.setCurrencyCode("USD");
         request.setLocaleCode("en-US");
@@ -93,6 +100,8 @@ public class PayPalCheckoutRequestUnitTest {
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
         request.setMerchantAccountId("merchant_account_id");
+        request.setEnablePayPalAppSwitch(appSwitchEnabled);
+        request.setUserAuthenticationEmail("test-email");
 
         ArrayList<PayPalLineItem> lineItems = new ArrayList<>();
         lineItems.add(new PayPalLineItem(PayPalLineItemKind.DEBIT, "An Item", "1", "1"));
@@ -110,6 +119,8 @@ public class PayPalCheckoutRequestUnitTest {
             result.getBillingAgreementDescription());
         assertTrue(result.isShippingAddressRequired());
         assertTrue(result.isShippingAddressEditable());
+        assertEquals(appSwitchEnabled, result.getEnablePayPalAppSwitch());
+        assertEquals("test-email", result.getUserAuthenticationEmail());
         assertEquals("Postal Address", result.getShippingAddressOverride().getRecipientName());
         assertEquals(PayPalPaymentIntent.SALE, result.getIntent());
         assertEquals(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN, result.getLandingPageType());
@@ -171,4 +182,5 @@ public class PayPalCheckoutRequestUnitTest {
 
         assertTrue(requestBody.contains("\"phone_number\":{\"country_code\":\"1\",\"national_number\":\"1231231234\"}"));
     }
+
 }