+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* PayPal
+    * Add PayPal App Switch for one time checkout flow (BETA)
+        * Add `enablePayPalAppSwitch` property to `PayPalCheckoutRequest` for App Switch support 
+        * Require `PayPalCheckoutRequest.userAuthenticationEmail` for App Switch support
+        * Require `PayPalClient.appLinkReturnUrl` for App Switch support
+        * **Note:** This feature is currently in beta and may change or be removed in future releases.
+
 ## 5.2.0 (2024-10-30)
 
 * GooglePay